+++ app/src/main/java/com/livechatinc/livechatwidgetexample/FullScreenWindowActivityExample.java
@@ -1,6 +1,7 @@
 package com.livechatinc.livechatwidgetexample;
 
-import android.content.Intent;
+import static android.view.View.GONE;
+
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
@@ -22,12 +23,6 @@ import com.livechatinc.inappchat.ChatWindowUtils;
 import com.livechatinc.inappchat.ChatWindowView;
 import com.livechatinc.inappchat.models.NewMessageModel;
 
-import static android.view.View.GONE;
-
-/**
- * Created by szymonjarosz on 26/07/2017.
- */
-
 public class FullScreenWindowActivityExample extends AppCompatActivity implements ChatWindowEventsListener {
     private FloatingActionButton startChatBtn;
     private ChatWindowView chatWindow;
@@ -41,25 +36,19 @@ public class FullScreenWindowActivityExample extends AppCompatActivity implement
         setContentView(R.layout.chat_window_launcher);
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
+
         chatWindow = ChatWindowUtils.createAndAttachChatWindowInstance(FullScreenWindowActivityExample.this);
-        chatWindow.setConfiguration((ChatWindowConfiguration) getIntent().getSerializableExtra("config"));
         chatWindow.setEventsListener(this);
-        chatWindow.initialize();
+        chatWindow.supportFileSharing(getActivityResultRegistry(), getLifecycle(), this);
+        chatWindow.init((ChatWindowConfiguration) getIntent().getSerializableExtra("config"));
+
         startChatBtn = findViewById(R.id.start_chat);
-        startChatBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                showChatWindow();
-            }
-        });
+        startChatBtn.setOnClickListener(view -> showChatWindow());
         chatBadgeTv = findViewById(R.id.chat_badge);
         clearSessionBtn = findViewById(R.id.clear_session_btn);
-        clearSessionBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                ChatWindowUtils.clearSession(view.getContext());
-                chatWindow.reload(false);
-            }
+        clearSessionBtn.setOnClickListener(view -> {
+            ChatWindowUtils.clearSession(view.getContext());
+            chatWindow.reload(false);
         });
     }
 
@@ -97,13 +86,13 @@ public class FullScreenWindowActivityExample extends AppCompatActivity implement
     }
 
     @Override
-    public void onWindowInitialized() {
-
+    public void onFilePickerActivityNotFound() {
+        Toast.makeText(this, "No file picker found", Toast.LENGTH_SHORT).show();
     }
 
     @Override
-    public void onStartFilePickerActivity(Intent intent, int requestCode) {
-        startActivityForResult(intent, requestCode);
+    public void onWindowInitialized() {
+
     }
 
     @Override
@@ -123,13 +112,6 @@ public class FullScreenWindowActivityExample extends AppCompatActivity implement
     public void onBackPressed() {
         if (!chatWindow.onBackPressed())
             super.onBackPressed();
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (!chatWindow.onActivityResult(requestCode, resultCode, data)) {
-            super.onActivityResult(requestCode, resultCode, data);
-        }
     }
 
     @Override