+++ inappchat/src/main/java/com/livechatinc/inappchat/FileChooserMode.java
@@ -1,0 +1,11 @@
+package com.livechatinc.inappchat;
+
+import android.os.Build;
+
+import androidx.annotation.RequiresApi;
+
+enum FileChooserMode {
+    SINGLE,
+    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    MULTIPLE
+}